requirements.txt
keras==2.6.0
keras_tuner==1.0.4
matplotlib==3.4.0
numpy==1.19.2
pandas==1.3.1
plotly==4.14.3
python_dateutil==2.8.2
scikit_learn==1.0.1
statsmodels==0.13.2
protobuf == 3.9.2
tensorflow==2.6.0
six == 1.15.0
